pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm", "versioneer[toml]==0.29"]
build-backend = "setuptools.build_meta"

[project]
name = "fisdat"
authors = [
	{ name = "William Waites" },
	{ name = "Meadhbh Moriarty" },
	{ name = "Duncan Guthrie" },
]
description = "SAVED: Data schemata validation and submission"
readme = "README.md"
requires-python = ">=3.5"
keywords = ["biological data", "semantic web"]
license = {text = "GPLv3"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
        "google-cloud-storage",
        "linkml",
        "linkml-runtime",
        "mypy",
        "versioneer"
]
dynamic = ["version"]

[project.scripts]
fisdat = "fisdat.cmd_dat:cli"
fisup = "fisdat.cmd_up:cli"

[tool.setuptools.dynamic]
version = {attr = "fisdat.__version__"}

[tool.setuptools]
packages = ["fisdat"]

[tool.versioneer]
VCS = "git"
style = "pep440"
versionfile_source = "fisdat/_version.py"
versionfile_build = "fisdat/_version.py"
tag_prefix = ""
parentdir_prefix = "fisdat"
